Fix timeout in amqp
The loop was missing a break after receiving ack,
so it would always wait the whole timeout time.

Waiting a whole second before checking for ack
again seems too long to me. This is even worse,
because the code execution is faster then the AMQP,
so it always sleeps at least once.